requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==1.12.0
sentence-transformers==2.5.1
psycopg2-binary==2.9.9
numpy==1.26.4
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
beautifulsoup4==4.12.3
selenium==4.16.0
tqdm==4.66.1
cryptography==44.0.2
transformers==4.37.2
tiktoken==0.6.0
PyMuPDF==1.23.8
Pillow==10.2.0
torch==2.2.1
faiss-cpu==1.7.4  # or faiss-gpu for GPU support
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
Flask-Session==0.6.0 